+++ experiments/test_chesapeakecvpr_models.py
@@ -12,7 +12,7 @@ import pytorch_lightning as pl
 import torch
 
 from torchgeo.datasets import ChesapeakeCVPRDataModule
-from torchgeo.trainers.segmentation import ChesapeakeCVPRSegmentationTask
+from torchgeo.trainers.chesapeake import ChesapeakeCVPRSegmentationTask
 
 ALL_TEST_SPLITS = [["de-val"], ["pa-test"], ["ny-test"], ["pa-test", "ny-test"]]
