+++ torchgeo/trainers/naipchesapeake.py
@@ -1,0 +1,66 @@
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License.
+
+"""Segmentation tasks."""
+
+from typing import Any, Dict
+
+import matplotlib.pyplot as plt
+import numpy as np
+from torch.utils.tensorboard import SummaryWriter  # type: ignore[attr-defined]
+
+from .segmentation import SemanticSegmentationTask
+
+
+# TODO: move this functionality into SemanticSegmentationTask and remove this class
+class NAIPChesapeakeSegmentationTask(SemanticSegmentationTask):
+    """LightningModule for training models on the NAIP and Chesapeake datasets.
+
+    .. deprecated: 0.1
+       Use :class:`SemanticSegmentationTask` instead.
+    """
+
+    def validation_step(  # type: ignore[override]
+        self, batch: Dict[str, Any], batch_idx: int
+    ) -> None:
+        """Validation step - reports average accuracy and average IoU.
+
+        Args:
+            batch: current batch
+            batch_idx: index of current batch
+        """
+        x = batch["image"]
+        y = batch["mask"]
+        y_hat = self.forward(x)
+        y_hat_hard = y_hat.argmax(dim=1)
+
+        loss = self.loss(y_hat, y)
+
+        # by default, the test and validation steps only log per *epoch*
+        self.log("val_loss", loss)
+        self.val_metrics(y_hat_hard, y)
+
+        if batch_idx < 10:
+            # Render the image, ground truth mask, and predicted mask for the first
+            # image in the batch
+            img = np.rollaxis(  # convert image to channels last format
+                batch["image"][0].cpu().numpy(), 0, 3
+            )
+            mask = batch["mask"][0].cpu().numpy()
+            pred = y_hat_hard[0].cpu().numpy()
+            fig, axs = plt.subplots(1, 3, figsize=(12, 4))
+            axs[0].imshow(img)
+            axs[0].axis("off")
+            axs[1].imshow(mask, vmin=0, vmax=4)
+            axs[1].axis("off")
+            axs[2].imshow(pred, vmin=0, vmax=4)
+            axs[2].axis("off")
+
+            # the SummaryWriter is a tensorboard object, see:
+            # https://pytorch.org/docs/stable/tensorboard.html#
+            summary_writer: SummaryWriter = self.logger.experiment
+            summary_writer.add_figure(
+                f"image/{batch_idx}", fig, global_step=self.global_step
+            )
+
+            plt.close()
