+++ torchgeo/trainers/segmentation.py
@@ -5,10 +5,6 @@
 
 from typing import Any, Dict, cast
 
-import kornia.augmentation as K
-import matplotlib
-import matplotlib.pyplot as plt
-import numpy as np
 import segmentation_models_pytorch as smp
 import torch
 import torch.nn as nn
@@ -16,32 +12,13 @@ from pytorch_lightning.core.lightning import LightningModule
 from torch import Tensor
 from torch.optim.lr_scheduler import ReduceLROnPlateau
 from torch.utils.data import DataLoader
-from torch.utils.tensorboard import SummaryWriter  # type: ignore[attr-defined]
 from torchmetrics import Accuracy, IoU, MetricCollection
 
-from ..datasets import Chesapeake7
 from ..models import FCN
 
 # https://github.com/pytorch/pytorch/issues/60979
 # https://github.com/pytorch/pytorch/pull/61045
 DataLoader.__module__ = "torch.utils.data"
-
-# TODO: move the color maps to a dataset object
-CMAP_7 = matplotlib.colors.ListedColormap(
-    [np.array(Chesapeake7.cmap[i]) / 255.0 for i in range(7)]
-)
-CMAP_5 = matplotlib.colors.ListedColormap(
-    np.array(
-        [
-            (0, 0, 0, 0),
-            (0, 197, 255, 255),
-            (38, 115, 0, 255),
-            (163, 255, 115, 255),
-            (156, 156, 156, 255),
-        ]
-    )
-    / 255.0
-)
 
 
 class SemanticSegmentationTask(LightningModule):
@@ -244,225 +221,3 @@ class SemanticSegmentationTask(LightningModule):
                 "monitor": "val_loss",
             },
         }
-
-
-# TODO: refactor any differences between these classes and SemanticSegmentationTask
-# so that these classes are no longer needed.
-class ChesapeakeCVPRSegmentationTask(SemanticSegmentationTask):
-    """LightningModule for training models on the Chesapeake CVPR Land Cover dataset.
-
-    .. deprecated: 0.1
-       Use :class:`SemanticSegmentationTask` instead.
-    """
-
-    def validation_step(  # type: ignore[override]
-        self, batch: Dict[str, Any], batch_idx: int
-    ) -> None:
-        """Validation step - reports average accuracy and average IoU.
-
-        Logs the first 10 validation samples to tensorboard as images with 3 subplots
-        showing the image, mask, and predictions.
-
-        Args:
-            batch: Current batch
-            batch_idx: Index of current batch
-        """
-        x = batch["image"]
-        y = batch["mask"]
-        y_hat = self.forward(x)
-        y_hat_hard = y_hat.argmax(dim=1)
-
-        loss = self.loss(y_hat, y)
-
-        self.log("val_loss", loss, on_step=False, on_epoch=True)
-        self.val_metrics(y_hat_hard, y)
-
-        if batch_idx < 10:
-            cmap = None
-            if self.hparams["num_classes"] == 5:
-                cmap = CMAP_5
-            else:
-                cmap = CMAP_7
-            # Render the image, ground truth mask, and predicted mask for the first
-            # image in the batch
-            img = np.rollaxis(  # convert image to channels last format
-                batch["image"][0].cpu().numpy(), 0, 3
-            )
-            mask = batch["mask"][0].cpu().numpy()
-            pred = y_hat_hard[0].cpu().numpy()
-            fig, axs = plt.subplots(1, 3, figsize=(12, 4))
-            axs[0].imshow(img[:, :, :3])
-            axs[0].axis("off")
-            axs[1].imshow(
-                mask,
-                vmin=0,
-                vmax=self.hparams["num_classes"] - 1,
-                cmap=cmap,
-                interpolation="none",
-            )
-            axs[1].axis("off")
-            axs[2].imshow(
-                pred,
-                vmin=0,
-                vmax=self.hparams["num_classes"] - 1,
-                cmap=cmap,
-                interpolation="none",
-            )
-            axs[2].axis("off")
-            plt.tight_layout()
-
-            # the SummaryWriter is a tensorboard object, see:
-            # https://pytorch.org/docs/stable/tensorboard.html#
-            summary_writer: SummaryWriter = self.logger.experiment
-            summary_writer.add_figure(
-                f"image/{batch_idx}", fig, global_step=self.global_step
-            )
-            plt.close()
-
-
-class LandCoverAISegmentationTask(SemanticSegmentationTask):
-    """LightningModule for training models on the Landcover.AI Dataset.
-
-    .. deprecated: 0.1
-       Use :class:`SemanticSegmentationTask` instead.
-    """
-
-    # TODO: move this to LandCoverAIDataModule
-    train_augmentations = K.AugmentationSequential(
-        K.RandomRotation(p=0.5, degrees=90),
-        K.RandomHorizontalFlip(p=0.5),
-        K.RandomVerticalFlip(p=0.5),
-        K.RandomSharpness(p=0.5),
-        K.ColorJitter(p=0.5, brightness=0.1, contrast=0.1, saturation=0.1, hue=0.1),
-        data_keys=["input", "mask"],
-    )
-
-    def training_step(  # type: ignore[override]
-        self, batch: Dict[str, Any], batch_idx: int
-    ) -> Tensor:
-        """Training step - reports average accuracy and average IoU.
-
-        Args:
-            batch: Current batch
-            batch_idx: Index of current batch
-
-        Returns:
-            training loss
-        """
-        x = batch["image"]
-        y = batch["mask"]
-        with torch.no_grad():
-            x, y = self.train_augmentations(x, y)
-        y = y.long().squeeze()
-
-        y_hat = self.forward(x)
-        y_hat_hard = y_hat.argmax(dim=1)
-
-        loss = self.loss(y_hat, y)
-
-        # by default, the train step logs every `log_every_n_steps` steps where
-        # `log_every_n_steps` is a parameter to the `Trainer` object
-        self.log("train_loss", loss, on_step=True, on_epoch=False)
-        self.train_metrics(y_hat_hard, y)
-
-        return cast(Tensor, loss)
-
-    def validation_step(  # type: ignore[override]
-        self, batch: Dict[str, Any], batch_idx: int
-    ) -> None:
-        """Validation step - reports average accuracy and average IoU.
-
-        Logs the first 10 validation samples to tensorboard as images with 3 subplots
-        showing the image, mask, and predictions.
-
-        Args:
-            batch: Current batch
-            batch_idx: Index of current batch
-        """
-        x = batch["image"]
-        y = batch["mask"].long().squeeze()
-        y_hat = self.forward(x)
-        y_hat_hard = y_hat.argmax(dim=1)
-
-        loss = self.loss(y_hat, y)
-
-        self.log("val_loss", loss, on_step=False, on_epoch=True)
-        self.val_metrics(y_hat_hard, y)
-
-        if batch_idx < 10 and self.hparams["verbose"]:
-            # Render the image, ground truth mask, and predicted mask for the first
-            # image in the batch
-            img = np.rollaxis(  # convert image to channels last format
-                x[0].cpu().numpy(), 0, 3
-            )
-            mask = y[0].cpu().numpy()
-            pred = y_hat_hard[0].cpu().numpy()
-            fig, axs = plt.subplots(1, 3, figsize=(12, 4))
-            axs[0].imshow(img)
-            axs[0].axis("off")
-            axs[1].imshow(mask, vmin=0, vmax=5)
-            axs[1].axis("off")
-            axs[2].imshow(pred, vmin=0, vmax=5)
-            axs[2].axis("off")
-
-            # the SummaryWriter is a tensorboard object, see:
-            # https://pytorch.org/docs/stable/tensorboard.html#
-            summary_writer: SummaryWriter = self.logger.experiment
-            summary_writer.add_figure(
-                f"image/{batch_idx}", fig, global_step=self.global_step
-            )
-
-            plt.close()
-
-
-class NAIPChesapeakeSegmentationTask(SemanticSegmentationTask):
-    """LightningModule for training models on the NAIP and Chesapeake datasets.
-
-    .. deprecated: 0.1
-       Use :class:`SemanticSegmentationTask` instead.
-    """
-
-    def validation_step(  # type: ignore[override]
-        self, batch: Dict[str, Any], batch_idx: int
-    ) -> None:
-        """Validation step - reports average accuracy and average IoU.
-
-        Args:
-            batch: current batch
-            batch_idx: index of current batch
-        """
-        x = batch["image"]
-        y = batch["mask"]
-        y_hat = self.forward(x)
-        y_hat_hard = y_hat.argmax(dim=1)
-
-        loss = self.loss(y_hat, y)
-
-        # by default, the test and validation steps only log per *epoch*
-        self.log("val_loss", loss)
-        self.val_metrics(y_hat_hard, y)
-
-        if batch_idx < 10:
-            # Render the image, ground truth mask, and predicted mask for the first
-            # image in the batch
-            img = np.rollaxis(  # convert image to channels last format
-                batch["image"][0].cpu().numpy(), 0, 3
-            )
-            mask = batch["mask"][0].cpu().numpy()
-            pred = y_hat_hard[0].cpu().numpy()
-            fig, axs = plt.subplots(1, 3, figsize=(12, 4))
-            axs[0].imshow(img)
-            axs[0].axis("off")
-            axs[1].imshow(mask, vmin=0, vmax=4)
-            axs[1].axis("off")
-            axs[2].imshow(pred, vmin=0, vmax=4)
-            axs[2].axis("off")
-
-            # the SummaryWriter is a tensorboard object, see:
-            # https://pytorch.org/docs/stable/tensorboard.html#
-            summary_writer: SummaryWriter = self.logger.experiment
-            summary_writer.add_figure(
-                f"image/{batch_idx}", fig, global_step=self.global_step
-            )
-
-            plt.close()
